Chore: 루틴 Cell과 View 분리
